requirements.txt
faiss-gpu
fastapi
numpy
opencv-python
pillow
pytesseract
sentence-transformers
torch
transformers
uvicorn
wordninja